% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Take the scalar product of two values}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{a numeric value}

\item{y}{a numeric value}
}
\value{
the product of x and y
}
\description{
Take the scalar product of two values
}
\examples{
multiply(3, 4)
multiply(4, 3)
}
